update based on review; cleanups and remove universal loaders
