+++ src/Slick/Pandoc.hs
@@ -9,7 +9,13 @@ module Slick.Pandoc
   , PandocWriter
   , loadUsing
   , loadUsing'
-  , loadEntity
+  , defaultMarkdownOptions
+  , defaultHtml5Options
+  , convert
+  , srcToURL
+  , destToSrc
+  , srcToDest
+  , flattenMeta
   ) where
 
 import           Control.Lens
@@ -22,14 +28,34 @@ import           Development.Shake.Classes
 import           Development.Shake.FilePath
 import           GHC.Generics               hiding (Meta)
 import           Text.Pandoc
+import           Text.Pandoc.Highlighting
 import           Text.Pandoc.Shared
-
-import           Slick.Build
 
 --------------------------------------------------------------------------------
 
 type PandocReader textType = textType -> PandocIO Pandoc
 type PandocWriter = Pandoc -> PandocIO T.Text
+
+-- | Reasonable options for reading a markdown file
+defaultMarkdownOptions :: ReaderOptions
+defaultMarkdownOptions =
+  def { readerExtensions = exts }
+  where
+    exts = mconcat
+     [ extensionsFromList
+       [ Ext_yaml_metadata_block
+       , Ext_fenced_code_attributes
+       , Ext_auto_identifiers
+       ]
+     , githubMarkdownExtensions
+     ]
+
+-- | Reasonable options for rendering to HTML
+defaultHtml5Options :: WriterOptions
+defaultHtml5Options =
+  def { writerHighlightStyle = Just tango
+      , writerExtensions     = writerExtensions def
+      }
 
 --------------------------------------------------------------------------------
 
@@ -112,29 +138,42 @@ loadUsing' :: (FromJSON a)
 loadUsing' reader writer text =
   loadUsing reader writer text >>= convert
 
--- | Generalized Entity Loader, wrapper in
---   own function to respolve EntityFilePath properly
-loadEntity :: FromJSON b
-           => ReaderOptions      -- ^ Pandoc reader options to specify extensions or other functionality
-           -> WriterOptions      -- ^ Pandoc writer options to modify output
-           -> FilePath           -- ^ Source directory
-           -> EntityFilePath a   -- ^ Path to entity ready for conversion, usually text file
-           -> Action b
-loadEntity rops wops inDir (EntityFilePath path) =
-  subloadEntity rops wops inDir path
+--------------------------------------------------------------------------------
 
--- | Helper functiona for generalized `loadEntity`
---   provides conversion from source to output file format
-subloadEntity :: FromJSON b
-              => ReaderOptions  -- ^ Pandoc reader options to specify extensions or other functionality
-              -> WriterOptions  -- ^ Pandoc writer options to modify output
-              -> FilePath       -- ^ Directory with files
-              -> FilePath       -- ^ Path to the file ready for conversion
-              -> Action b
-subloadEntity rops wops dir entityPath = do
-  let srcPath = destToSrc dir entityPath -<.> "md"
-  entityData <- readFile' srcPath >>= markdownToHTML rops wops . T.pack
-  let entityURL = T.pack  . srcToURL $ entityPath
-      withURL = _Object . at "url"     ?~ String entityURL
-      withSrc = _Object . at "srcPath" ?~ String (T.pack srcPath)
-  convert . withSrc . withURL $ entityData
+-- | Convert 'build' filepaths into source file filepaths
+destToSrc :: FilePath  -- ^ input directory "site"
+          -> FilePath
+          -> FilePath
+destToSrc i p = i </> dropDirectory1 p
+
+-- | Convert source filepaths into build filepaths
+srcToDest :: FilePath  -- ^ input directory, "dist"
+          -> FilePath  -- ^
+          -> FilePath  -- ^
+srcToDest i p = i </> dropDirectory1 p
+
+-- | Convert a source file path into a URL
+srcToURL :: FilePath -> String
+srcToURL = ("/" ++) . dropDirectory1 . (-<.> ".html")
+
+--------------------------------------------------------------------------------
+
+-- | Attempt to convert between two JSON serializable objects (or 'Value's).
+--   Failure to deserialize fails the Shake build.
+convert :: (FromJSON a, ToJSON a, FromJSON b) => a -> Action b
+convert a = case fromJSON (toJSON a) of
+  Success r   -> pure r
+  Error   err -> fail $ "json conversion error:" ++ err
+
+-- | Flatten a Pandoc 'Meta' into a well-structured JSON object, rendering Pandoc
+--   text objects into plain strings along the way.
+flattenMeta :: Meta -> Value
+flattenMeta (Meta meta) = toJSON $ fmap go meta
+ where
+  go :: MetaValue -> Value
+  go (MetaMap     m) = toJSON $ fmap go m
+  go (MetaList    m) = toJSONList $ fmap go m
+  go (MetaBool    m) = toJSON m
+  go (MetaString  m) = toJSON m
+  go (MetaInlines m) = toJSON $ stringify m
+  go (MetaBlocks  m) = toJSON $ stringify m
